requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.1
urllib3==2.0.2
vonage==3.5.1
wrapt==1.15.0
yarg==0.1.9
yarl==1.9.2
